requirements.txt
python-dateutil
pytz
feedparser
requests
markdown2
html2text
newspaper3k
python-dotenv
fastapi
starlette
fastapi-cache2
psycopg2
apscheduler
uvicorn
lxml[html_clean]
jinja2
python-multipart
psycopg2-binary
python-telegram-bot
openai
nest_asyncio